requirements.txt
Flask==3.0.3
torch==2.6.0
numpy
Pillow
serverless-wsgi
gunicorn
pytest
torchvision==2.6.0